+++ pkg/services/object/get/res.go
@@ -1,9 +1,0 @@
-package getsvc
-
-type RangeHashRes struct {
-	hashes [][]byte
-}
-
-func (r *RangeHashRes) Hashes() [][]byte {
-	return r.hashes
-}