+++ pkg/services/object/server.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"context"
 	"crypto/ecdsa"
-	"crypto/sha256"
 	"encoding/base64"
 	"encoding/binary"
 	"errors"
@@ -48,7 +47,6 @@ import (
 	"github.com/nspcc-dev/neofs-sdk-go/stat"
 	"github.com/nspcc-dev/neofs-sdk-go/user"
 	"github.com/nspcc-dev/neofs-sdk-go/version"
-	"github.com/nspcc-dev/tzhash/tz"
 	"github.com/panjf2000/ants/v2"
 	"google.golang.org/grpc"
 	grpccodes "google.golang.org/grpc/codes"
@@ -65,7 +63,6 @@ type Handlers interface {
 	Head(context.Context, getsvc.HeadPrm) error
 	Delete(context.Context, deletesvc.Prm) error
 	GetRange(context.Context, getsvc.RangePrm) error
-	GetRangeHash(context.Context, getsvc.RangeHashPrm) (*getsvc.RangeHashRes, error)
 }
 
 // Various NeoFS protocol status codes.
@@ -83,7 +80,6 @@ const (
 //   - [stat.MethodObjectDelete]
 //   - [stat.MethodObjectSearch]
 //   - [stat.MethodObjectRange]
-//   - [stat.MethodObjectHash]
 type MetricCollector interface {
 	// HandleOpExecResult handles measured execution results of the given op.
 	HandleOpExecResult(_ stat.Method, success bool, _ time.Duration)
@@ -160,7 +156,6 @@ type ACLInfoExtractor interface {
 	PutRequestToInfo(*protoobject.PutRequest) (aclsvc.RequestInfo, user.ID, error)
 	DeleteRequestToInfo(*protoobject.DeleteRequest) (aclsvc.RequestInfo, error)
 	HeadRequestToInfo(*protoobject.HeadRequest) (aclsvc.RequestInfo, error)
-	HashRequestToInfo(*protoobject.GetRangeHashRequest) (aclsvc.RequestInfo, error)
 	GetRequestToInfo(*protoobject.GetRequest) (aclsvc.RequestInfo, error)
 	RangeRequestToInfo(*protoobject.GetRangeRequest) (aclsvc.RequestInfo, error)
 	SearchV2RequestToInfo(*protoobject.SearchV2Request) (aclsvc.RequestInfo, error)
@@ -825,185 +820,9 @@ func convertHeadPrm(signer ecdsa.PrivateKey, cnr container.Container, req *proto
 	return p, nil
 }
 
-func (s *Server) signHashResponse(resp *protoobject.GetRangeHashResponse, req *protoobject.GetRangeHashRequest) *protoobject.GetRangeHashResponse {
-	resp.VerifyHeader = util.SignResponseIfNeeded(&s.signer, resp, req)
-	return resp
-}
-
-func (s *Server) makeStatusHashResponse(err error, req *protoobject.GetRangeHashRequest) *protoobject.GetRangeHashResponse {
-	return s.signHashResponse(&protoobject.GetRangeHashResponse{
-		MetaHeader: s.makeResponseMetaHeader(util.ToStatus(err)),
-	}, req)
-}
-
-// GetRangeHash converts gRPC GetRangeHashRequest message and passes it to internal Object service.
-func (s *Server) GetRangeHash(ctx context.Context, req *protoobject.GetRangeHashRequest) (*protoobject.GetRangeHashResponse, error) {
-	var (
-		err error
-		t   = time.Now()
-	)
-	defer func() { s.pushOpExecResult(stat.MethodObjectHash, err, t) }()
-	if err = icrypto.VerifyRequestSignaturesN3(req, s.fsChain); err != nil {
-		return s.makeStatusHashResponse(err, req), nil
-	}
-
-	if s.fsChain.LocalNodeUnderMaintenance() {
-		return s.makeStatusHashResponse(apistatus.ErrNodeUnderMaintenance, req), nil
-	}
-
-	reqInfo, err := s.reqInfoProc.HashRequestToInfo(req)
-	if err != nil {
-		if !errors.Is(err, apistatus.Error) {
-			var bad = new(apistatus.BadRequest)
-			bad.SetMessage(err.Error())
-			err = bad // defer
-		}
-		return s.makeStatusHashResponse(err, req), nil
-	}
-	if !s.aclChecker.CheckBasicACL(reqInfo) {
-		err = basicACLErr(reqInfo) // needed for defer
-		return s.makeStatusHashResponse(err, req), nil
-	}
-	err = s.aclChecker.CheckEACL(req, reqInfo)
-	if err != nil && !errors.Is(err, aclsvc.ErrNotMatched) { // Not matched -> follow basic ACL.
-		err = eACLErr(reqInfo, err) // needed for defer
-		return s.makeStatusHashResponse(err, req), nil
-	}
-
-	p, err := convertHashPrm(s.signer, reqInfo.Container, s.storage, req)
-	if err != nil {
-		if !errors.Is(err, apistatus.Error) {
-			var bad = new(apistatus.BadRequest)
-			bad.SetMessage(err.Error())
-			err = bad // defer
-		}
-		return s.makeStatusHashResponse(err, req), nil
-	}
-	res, err := s.handlers.GetRangeHash(ctx, p)
-	if err != nil {
-		return s.makeStatusHashResponse(err, req), nil
-	}
-
-	return s.signHashResponse(&protoobject.GetRangeHashResponse{
-		Body: &protoobject.GetRangeHashResponse_Body{
-			Type:     req.Body.Type,
-			HashList: res.Hashes(),
-		}}, req), nil
-}
-
-// converts original request into parameters accepted by the internal handler.
-func convertHashPrm(signer ecdsa.PrivateKey, cnr container.Container, ss sessions, req *protoobject.GetRangeHashRequest) (getsvc.RangeHashPrm, error) {
-	body := req.GetBody()
-	ma := body.GetAddress()
-	if ma == nil { // includes nil body
-		return getsvc.RangeHashPrm{}, errors.New("missing object address")
-	}
-
-	var addr oid.Address
-	if err := addr.FromProtoMessage(ma); err != nil {
-		return getsvc.RangeHashPrm{}, fmt.Errorf("invalid object address: %w", err)
-	}
-
-	cp, err := objutil.CommonPrmFromRequest(req)
-	if err != nil {
-		return getsvc.RangeHashPrm{}, err
-	}
-
-	var p getsvc.RangeHashPrm
-
-	switch t := body.GetType(); t {
-	default:
-		return getsvc.RangeHashPrm{}, fmt.Errorf("unknown checksum type %v", t)
-	case refs.ChecksumType_SHA256:
-		p.SetHashGenerator(sha256.New)
-	case refs.ChecksumType_TZ:
-		p.SetHashGenerator(tz.New)
-	}
-
-	if tokV2 := cp.SessionTokenV2(); tokV2 != nil {
-		signerKey, err := ss.GetSessionV2PrivateKey(tokV2.Subjects())
-		if err != nil {
-			if !errors.Is(err, apistatus.ErrSessionTokenNotFound) {
-				return getsvc.RangeHashPrm{}, fmt.Errorf("fetching session v2 key: %w", err)
-			}
-			cp.ForgetTokens()
-			signerKey = signer
-		}
-		p.WithCachedSignerKey(&signerKey)
-	} else if tok := cp.SessionToken(); tok != nil {
-		authUser, err := tok.AuthUser()
-		if err != nil {
-			return getsvc.RangeHashPrm{}, fmt.Errorf("getting auth user from token: %w", err)
-		}
-		signerKey, err := ss.GetSessionPrivateKey(authUser)
-		if err != nil {
-			if !errors.Is(err, apistatus.ErrSessionTokenNotFound) {
-				return getsvc.RangeHashPrm{}, fmt.Errorf("fetching session key: %w", err)
-			}
-			cp.ForgetTokens()
-			signerKey = signer
-		}
-		p.WithCachedSignerKey(&signerKey)
-	}
-
-	mr := body.GetRanges()
-	rngs := make([]object.Range, len(mr))
-	for i := range mr {
-		rngs[i].SetOffset(mr[i].Offset)
-		rngs[i].SetLength(mr[i].Length)
-	}
-
-	p.SetCommonParameters(cp)
-	p.WithAddress(addr)
-	p.WithContainer(cnr)
-	p.SetRangeList(rngs)
-	p.SetSalt(body.GetSalt())
-
-	if cp.LocalOnly() {
-		return p, nil
-	}
-
-	var onceResign sync.Once
-	meta := req.GetMetaHeader()
-	if meta == nil {
-		return getsvc.RangeHashPrm{}, errors.New("missing meta header")
-	}
-	p.SetRangeHashRequestForwarder(func(ctx context.Context, c client.MultiAddressClient) ([][]byte, error) {
-		var err error
-		onceResign.Do(func() {
-			req.MetaHeader = &protosession.RequestMetaHeader{
-				// TODO: #1165 think how to set the other fields
-				Version: newCurrentProtoVersionMessage(),
-				Ttl:     meta.GetTtl() - 1,
-				Origin:  meta,
-			}
-			req.VerifyHeader, err = neofscrypto.SignRequestWithBuffer(neofsecdsa.Signer(signer), req, nil)
-		})
-		if err != nil {
-			return nil, err
-		}
-
-		var hs [][]byte
-		return hs, c.ForAnyGRPCConn(ctx, func(ctx context.Context, conn *grpc.ClientConn) error {
-			var err error
-			hs, err = getHashesFromRemoteNode(ctx, conn, req)
-			return err // TODO: log error
-		})
-	})
-	return p, nil
-}
-
-func getHashesFromRemoteNode(ctx context.Context, conn *grpc.ClientConn, req *protoobject.GetRangeHashRequest) ([][]byte, error) {
-	resp, err := protoobject.NewObjectServiceClient(conn).GetRangeHash(ctx, req)
-	if err != nil {
-		return nil, fmt.Errorf("GetRangeHash rpc failure: %w", err)
-	}
-
-	if err := checkStatus(resp.GetMetaHeader().GetStatus()); err != nil {
-		return nil, err
-	}
-	// TODO: verify number of hashes
-	return resp.GetBody().GetHashList(), nil
+// GetRangeHash is deprecated and no longer supported by the node.
+func (s *Server) GetRangeHash(_ context.Context, _ *protoobject.GetRangeHashRequest) (*protoobject.GetRangeHashResponse, error) {
+	return nil, grpcstatus.Error(grpccodes.Unimplemented, "no longer supported")
 }
 
 func (s *Server) sendGetResponse(stream protoobject.ObjectService_GetServer, resp *protoobject.GetResponse, sign bool) error {
@@ -1036,9 +855,14 @@ type getStream struct {
 
 	recheckEACL  bool
 	signResponse bool
+	payloadOnly  bool
 }
 
-func (s *getStream) WriteHeader(hdr *object.Object) error {
+func (s *getStream) ValidateHeader(hdr *object.Object) error {
+	if !s.recheckEACL {
+		return nil
+	}
+
 	mo := hdr.ProtoMessage()
 	resp := &protoobject.GetResponse{
 		Body: &protoobject.GetResponse_Body{
@@ -1049,11 +873,31 @@ func (s *getStream) WriteHeader(hdr *object.Object) error {
 			}},
 		},
 	}
-	if s.recheckEACL {
-		err := s.srv.aclChecker.CheckEACL(resp, s.reqInfo)
-		if err != nil && !errors.Is(err, aclsvc.ErrNotMatched) { // Not matched -> follow basic ACL.
-			return eACLErr(s.reqInfo, err)
-		}
+
+	err := s.srv.aclChecker.CheckEACL(resp, s.reqInfo)
+	if err != nil && !errors.Is(err, aclsvc.ErrNotMatched) { // Not matched -> follow basic ACL.
+		return eACLErr(s.reqInfo, err)
+	}
+	return nil
+}
+
+func (s *getStream) WriteHeader(hdr *object.Object) error {
+	if err := s.ValidateHeader(hdr); err != nil {
+		return err
+	}
+	if s.payloadOnly {
+		return nil
+	}
+
+	mo := hdr.ProtoMessage()
+	resp := &protoobject.GetResponse{
+		Body: &protoobject.GetResponse_Body{
+			ObjectPart: &protoobject.GetResponse_Body_Init_{Init: &protoobject.GetResponse_Body_Init{
+				ObjectId:  mo.ObjectId,
+				Signature: mo.Signature,
+				Header:    mo.Header,
+			}},
+		},
 	}
 	return s.srv.sendGetResponse(s.base, resp, s.signResponse)
 }
@@ -1121,6 +965,7 @@ func (s *Server) Get(req *protoobject.GetRequest, gStream protoobject.ObjectServ
 		reqInfo:      reqInfo,
 		recheckEACL:  recheckEACL,
 		signResponse: needSignResp,
+		payloadOnly:  req.GetBody().GetPayloadOnly(),
 	})
 	if err != nil {
 		if !errors.Is(err, apistatus.Error) {
@@ -1135,8 +980,12 @@ func (s *Server) Get(req *protoobject.GetRequest, gStream protoobject.ObjectServ
 	// We could acquire ~256K buffer (like for chunks) if storage would try to read it full.
 	// Then small objects would fit into a single buffer, and for large ones it'd be possible to
 	// encode the first chunk response using the heading buffer.
-	hdrRespBuf, hdrBuf := getBufferForHeadResponse()
-	defer hdrRespBuf.Free()
+	var hdrRespBuf *iprotobuf.MemBuffer
+	var hdrBuf []byte
+	if p.Range() == nil && !p.PayloadOnly() {
+		hdrRespBuf, hdrBuf = getBufferForHeadResponse()
+		defer hdrRespBuf.Free()
+	}
 
 	hdrLen := -1
 	var stream io.ReadCloser
@@ -1146,7 +995,9 @@ func (s *Server) Get(req *protoobject.GetRequest, gStream protoobject.ObjectServ
 		}
 	}()
 
-	p.WithBuffer(hdrBuf, func(ln int, s io.ReadCloser) { hdrLen, stream = ln, s })
+	if hdrBuf != nil {
+		p.WithBuffer(hdrBuf, func(ln int, s io.ReadCloser) { hdrLen, stream = ln, s })
+	}
 
 	err = s.handlers.Get(gStream.Context(), p)
 	if err != nil {
@@ -1296,6 +1147,18 @@ func convertGetPrm(signer ecdsa.PrivateKey, cnr container.Container, req *protoo
 		return getsvc.Prm{}, fmt.Errorf("invalid object address: %w", err)
 	}
 
+	rng := body.GetRange()
+	if rng != nil {
+		rln := rng.GetLength()
+		if rln == 0 {
+			if rng.GetOffset() != 0 {
+				return getsvc.Prm{}, errors.New("zero range length")
+			}
+		} else if rng.GetOffset()+rln <= rng.GetOffset() {
+			return getsvc.Prm{}, errors.New("range overflow")
+		}
+	}
+
 	cp, err := objutil.CommonPrmFromRequest(req)
 	if err != nil {
 		return getsvc.Prm{}, err
@@ -1307,6 +1170,15 @@ func convertGetPrm(signer ecdsa.PrivateKey, cnr container.Container, req *protoo
 	p.WithContainer(cnr)
 	p.WithRawFlag(body.Raw)
 	p.SetObjectWriter(stream)
+	if rng != nil {
+		var objRng object.Range
+		objRng.SetOffset(rng.GetOffset())
+		objRng.SetLength(rng.GetLength())
+		p.SetRange(&objRng)
+	}
+	if body.GetPayloadOnly() {
+		p.MarkPayloadOnly()
+	}
 	if cp.LocalOnly() {
 		return p, nil
 	}
@@ -1318,14 +1190,18 @@ func convertGetPrm(signer ecdsa.PrivateKey, cnr container.Container, req *protoo
 	}
 
 	proxyCtx := getProxyContext{
-		req:        req,
-		reqOID:     addr.Object(),
-		respStream: stream,
+		req:          req,
+		reqOID:       addr.Object(),
+		respStream:   stream,
+		suppressInit: body.GetPayloadOnly(),
 	}
 
 	p.SetRequestForwarder(func(ctx context.Context, c client.MultiAddressClient) error {
 		var err error
 		onceResign.Do(func() {
+			if proxyCtx.suppressInit {
+				req.Body.PayloadOnly = false
+			}
 			req.MetaHeader = &protosession.RequestMetaHeader{
 				// TODO: #1165 think how to set the other fields
 				Ttl:    meta.GetTtl() - 1,
@@ -1345,9 +1221,10 @@ func convertGetPrm(signer ecdsa.PrivateKey, cnr container.Container, req *protoo
 }
 
 type getProxyContext struct {
-	req        *protoobject.GetRequest
-	reqOID     oid.ID
-	respStream *getStream
+	req          *protoobject.GetRequest
+	reqOID       oid.ID
+	respStream   *getStream
+	suppressInit bool
 
 	onceHdr sync.Once