+++ CHANGELOG.md
@@ -39,14 +39,17 @@ Changelog for NeoFS Node
 - Optimized netmap caching in node (#3966)
 - Store in metabase associated object ID in bytes instead of Base58 (#3971)
 - Optimized local RANGE request execution (#3967)
+- GET now supports payload ranges (#3991)
 
 ### Removed
 - `policer.max_workers` configuration (#3920)
 - Deprecated `Search` method support from storage nodes (#3931)
 - Internal conversions of node addresses received from network map (#3981)
+- `GetRangeHash` method support from storage nodes and related `object hash` CLI command (#3991)
 
 ### Updated
 - NeoGo dependency to 0.119.0 (#3993)
+- `github.com/nspcc-dev/neofs-sdk-go` module to `v1.0.0-rc.18.0.20260513135441-5c10a9626760` (#3991)
 
 ### Updating from v0.52.0
 Drop `policer.max_workers` configuration, it's no-op since 0.52.0.
@@ -59,6 +62,12 @@ migrate to `Searchv2` if needed.
 
 Metabase will migrate to version 11 with this release to update object
 counters, this can take a while for shards with high object numbers.
+
+Storage nodes no longer implement deprecated `GetRangeHash` method and
+`neofs-cli object hash` command has been removed.
+
+GET now supports payload ranges. Migrate from deprecated RANGE requests to GET
+with range parameters.
 
 ## [0.52.0] - 2026-03-27 - Woodo